backend/requirements.txt
# Web framework
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Metadata engines
Pillow==10.4.0
piexif==1.1.3
pillow-heif==0.16.0
PyExifTool==0.5.6
pypdf==4.2.0
python-docx==1.1.2

# MIME detection (requires libmagic system library)
python-magic==0.4.27

# Cache + rate limiting
cachetools==5.3.3
flask-limiter==3.7.0

# Dev / testing
pytest==8.2.2
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.4
